+++ source/databricks/calculation_engine/tests/datamigration/migration/test_migration.py
@@ -13,16 +13,8 @@
 # limitations under the License.
 from unittest.mock import Mock
 
-import spark_sql_migrations.schema_migration_pipeline as schema_migration_pipeline
 from pyspark.sql import SparkSession
 from pyspark.sql.types import StructType, StructField
-from package.infrastructure.paths import (
-    OUTPUT_DATABASE_NAME,
-    INPUT_DATABASE_NAME,
-    BASIS_DATA_DATABASE_NAME,
-    SETTLEMENT_REPORT_DATABASE_NAME,
-    EdiResults,
-)
 import package.datamigration.migration as sut
 import package.datamigration.schema_config as schema_config
 import tests.helpers.mock_helper as mock_helper
@@ -100,53 +92,3 @@ def test__migrate__when_schema_migration_scripts_are_executed__compare_result_wi
             ), f"Table {table.name} is not in the schema config"
             actual_table = spark.table(f"{db.name}.{table.name}")
             assert actual_table.schema == table_config.schema
-
-
-def test__schema_config__when_current_state_script_files_are_executed(
-    mocker: Mock, spark: SparkSession
-) -> None:
-    # Arrange
-    storage_account = "storage_account_4"
-    mocker.patch.object(
-        sut.paths,
-        sut.paths.get_storage_account_url.__name__,
-        side_effect=mock_helper.base_path_helper,
-    )
-
-    mocker.patch.object(
-        sut.env_vars,
-        sut.env_vars.get_storage_account_name.__name__,
-        return_value=storage_account,
-    )
-
-    mocker.patch.object(
-        sut.env_vars,
-        sut.env_vars.get_calculation_input_folder_name.__name__,
-        return_value=storage_account,
-    )
-
-    mocker.patch.object(
-        sut.paths,
-        sut.paths.get_spark_sql_migrations_path.__name__,
-        return_value=storage_account,
-    )
-
-    mocker.patch.object(
-        sut.paths,
-        sut.paths.get_container_root_path.__name__,
-        return_value=storage_account,
-    )
-
-    spark_helper.reset_spark_catalog(spark)
-    spark_sql_migration_helper.migrate_with_current_state(spark)
-
-    # Act
-    sut.migrate_data_lake()
-
-    # Assert
-    schemas = spark.catalog.listDatabases()
-    for schema in schema_config.schema_config:
-        assert schema.name in [schema.name for schema in schemas]
-        for table in schema.tables:
-            actual_table = spark.table(f"{schema.name}.{table.name}")
-            assert actual_table.schema == table.schema
